requirements.txt
# dependencies
beautifulsoup4
colorama
duckduckgo_search
md2pdf
playwright
openai
python-dotenv
pyyaml
uvicorn
pydantic
fastapi
python-multipart
markdown
# langchain>=0.2,<0.3
# langchain_community>=0.2,<0.3
# langchain-openai>=0.1,<0.2
# langchain-google-genai>=1,<2
# langchain-groq>=0.1,<0.2
# langchain_anthropic>=0.1,<0.2
# langchain_mistralai>=0.1,<0.2
# langchain_huggingface>=0.0.1,<0.1
# langchain_together>=0.1,<0.2
# langchain_cohere
# langchain_google_vertexai
# langchain_fireworks
# langchain_aws
tiktoken
tavily-python
gpt-researcher
arxiv
PyMuPDF
requests
jinja2
aiofiles
newspaper3k
SQLAlchemy
mistune
python-docx
htmldocx
lxml[html_clean]
websockets
unstructured
pandas
json_repair
exa_py

# Additional dependencies from multi_agents/requirements.txt
langgraph-cli
langgraph
weasyprint
json5

# uncomment for testing
# pytest
# pytest-asyncio